Complete initial RMap component installation, improve scripts
The initial code for installing the RMap components is now complete.
Some bugs were fixed in the NOID and API portions of the script, and the
Account Manager portion was added.  All scripts were made to execute
only once during any run.  The use of firewalld was replaced with calls
to iptables in the Tomcat script.  Assorted small bugs fixes and cleanup
is included.