Continuing improvements for RMap installer
This submission contains the following changes:
* Adds a script to generate a signed SSL certificate and import it into
a keystore used by Tomcat to provide HTTPS access to the site.  Note
that this isn't yet tied in to the GraphDB site, nor can a user's own
certificate be imported.
* Roll back the version of Tomcat we are using to fix a bug that broke
the API.
* Fixes for the MySQL installer so it behaves better when run a second
time ("upgrade" installs).
* Back up and restore Tomcat data when running an upgrade install for
it.  The data presently consists of the keystore, but may be extended to
include installed webapps, like the RMap components.