A collection of improvements for the install scripts
This submission contains the following improvements for the installer:
* Move the Tomcat location from the root URL to /tomcat.
* Change some properties to try to get MariaDB working.  Unfortunately,
these changes and other attempted fixes have not yet resulted in RMap
being able to communicate with MariaDB, and more work is needed.
* Rename a script, some variables and some comments from MySQL to
MariaDB.
* Switch to the latest 8.0.X build of Tomcat (45) and confirm that it
works with RMap.  The 8.5.X versions do not allow the API to work
correctly for some reason.
* Improve the error message if a script is not run as sudo (now shows
how to invoke the script).
* Don't reinstall Tomcat if its version hasn't changed.  This prevents
wasted time and having the RMap components and the keystore from being
removed needlessly from the installed Tomcat.
* Several other small fixes.