Improve upgrade install behavior and fix certificate generation
This submission contains a mixture of changes in three areas: improved
upgrade behavior, corrected certificate generation based on a
configuration file, and script output/logging.

When the install scripts are run for a second or later time, they are
considered to be running in "upgrade" mode.  Previously, this involved
saving a backup copy of some installed components and then reinstalling
them.  A recent submission changed this behavior for Tomcat, only
re-installing if the version number had changed.  This submission does
the same for the GraphDB and NOID components.  It also checks for
updates to the supporting Linux modules and upgrades those that need it.

The script that previously generated an SSL certificate and installed it
in Tomcat's keystore has been broken down into two scripts, one for
generation and one for installation.  This will allow a user-provided
certificate file to be installed more easily.  The behavior of these
scripts is controlled by a new configuration script where passwords and
the site domain name can be specified.  Eventually, this script will
change to read these values from the RMap properties file (some new
properties need to be added to that file first).

The certificate generation using Let's Encrypt became broken when we
changed the site structure so that RMap was at the root and Tomcat's
folder was in /tomcat.  It turns out that the cerbot's method of
temporarily placing content in the root of the site and then looking for
it via HTTP doesn't work in the RMap webapp.  To fix this, the script
now moves the RMap "ROOT" folder (or creates one if the RMap component
is not yet installed) to allow the certbot to work correctly, and then
cleans up afterwards.

The console and logging output of the installer scripts was enhanced so
that each component prints an introduction line in white, and the
component intro and outro lines are shown in bold.  this allows the log
file (which contains additional white text from the command outputs) to
be visually parsed more easily.

The README file was also updated to describe the new behavior and usage.