Finish getting all fundamental behavior working correctly
With this submission, an install on a Google cloud instance now runs
smoothly and allows all initial operations to be performed without
problems.  The specific changes here are:
* Disable the firewalld service and add some settings to iptables to get
the Tomcat site to come up correctly after installation and after the
instance is rebooted.
* Add to and correct the configuration properties so that the RMap API
works after installation.
* Add some properties to the configuration to allow users to set the
user name and password for MariaDB and GraphDB before the installation.
These values are included in the rmap.properties file but are also set
within those databases during the run of the installer.
* Replace the previous tests with a more robust one to check when the
RMap account manager and API components are fully installed and "up" in
the site.
* Rename some variables and add some documentation to the scripts.
* Update the README file with the latest directions and information.