Allow su'd user to own files, fixes for independent GraphDB install.
Previously, if someone who was logged in as root then su'd to another
user, like 'rmap', and then ran the scripts, any files that were created
would not belong to 'rmap' as desired, but would belong to root.  Only
direct logins as 'rmap' would give the proper ownership.  This is fixed
by using a different command to determine the desired user name.

Testing to make sure that the GraphDB component could be installed on a
separate server turned up some issues with the GraphDB install script,
which are fixed here.  The script did not independently read the
configuration information and so when run by itself was not able to
create the initial user in the new GraphDB.  Some initialization
commands were also failing (at least on the Red Cloud instance I was
using) because the initial admin/root credentials were not being
supplied.  Some cleanup also extracts the JSON for the user creation
into a separate bit of code so that the three GraphDB API calls can be
more similar to each other.