Changes to support Amazon Cloud
The installer scripts ran fine on AWS with a few exceptions, which are
fixed by these changes.

The minimal CentOS 7 that is likely to be used for RMap servers does not
come with the firewalld service installed and a command to disable it
failed.  The script now checks for the presence of the service before
trying to disable it.

The installation of perl-BerkeleyDB in support of the RMap components
depends on the installation of epel-release, which is also missing from
the minimal CentOS on AWS.  This module is now checked/installed before
perl-BerkeleyDB.

It should also be noted that the installation will only succeed on AWS
instances having at least 2GB of RAM (size "T2 small" and above).  In
practice the servers will probably need more memory than that to
function once they are active and accumulating data.