Initial implementation of backup/restore for GraphDB content
This submission includes initial versions of two scripts that operate on
the GraphDB triplestore contents.  The backup_triplestore.sh script
grabs a snapshot of the statements in the DB and saves them to a
date-time-stamped file in a "backups" folder in the GraphDB installation
folder.  The restore_triplestore.sh script uploads a named snapshot file
to GraphDB, replacing any previous content that was there.

After these scripts are tested, it is expected that they will be
augmented to provide a better user interface and possibly add
functionality.  At that time, they should be documented in the RMap
Wiki.