Manage firewall for DB components on separate servers
During testing it was noticed that when a GraphDB server was installed
on a computer other than the one where Tomcat was installed, the
firewall settings were not correct after the computer was rebooted.  The
Tomcat installation creates some firewall settings that are reloaded
after reboot, but these were missing from the separate GraphDB computer.
This submission performs a version of that firewall setup for both
GraphDB and MariaDB when they are installed on a separate computer.