Improvements for backup/restore scripts
The initial implementation of the backup and restore scripts has been
augmented with a few configuration parameters to allow users to
customize its behavior.  Users can now specify the folder on the current
machine where the downloaded file should be saved, the URL to the
GraphDB web app (for custom installations) and the MIME type and file
extension to use for the download.  The same download folder and URL are
used by the script that restores a database.