Improvements for database restore scripts
Several improvements are made to the script that uploads data to the
running GraphDB:
* A new configuration parameter allows the heap size for GraphDB to be
specified.  Larger heap sizes are needed if large files are being
uploaded to the DB.
* The uploading is performed with a chunk size of 200,000 statements so
that the memory demand is not as great for large files.
* Instead of copying the file that is being uploaded into the required
folder, a soft link pointing to the file is placed in the folder.

Additionally, a second "offline" database restore script has been
created.  It uses the LoadRDF functionality from GraphDB to upload
statements more quickly, but with the GraphDB process stopped.

Both scripts could be further improved by splitting very large statement
files into more manageable sizes before uploading, but this action can
also be performed outside of the scripts.  Such functionality would be
highly dependent on the heap size of the GraphDB process.